requirements.txt
flask
flask_restx
python-dotenv
kink
beautifulsoup4
pandas
numpy
httpx
lxml
snapshottest